Add Streamz pane with ReplacementPane base class
